producer/requirements.txt
streamlit
requests
kafka-python
pandas 
google-genai